feat: add the measured capture (TCG2 OVMF, PCR 11 fixture, triangle) (#64)
* feat: add the measured capture (TCG2 OVMF, PCR 11 fixture, triangle)

Close the M4 measured-boot proof the TCG2 blocker held back
(M4-028c, ADR-0024). The blocker was re-verified empirically: a
traced swtpm shows zero firmware TPM2_Extend traffic under Fedora's
OVMF, a full guest boot leaves PCRs 0-9 and 11 at zero with no event
log (the observed PCR 10 activity is the kernel's IMA plus its TPM2
HMAC-session support, not firmware), and compressed firmware volumes
mean strings alone can never prove either side.

The capture: image/fetch-ovmf-tcg2.sh pins Ubuntu's hash-verified
TCG2-enabled ovmf-generic; image/boot-capture.sh builds a second
TEST-ONLY-signed UKI with the capture init embedded (sd-stub measures
it into PCR 11 through the firmware TCG2 protocol), boots it with an
IDE evidence disk (the guest ships no virtio modules; ata_piix is
builtin), and the guest exports the binary TCG2 event log, the live
PCR values, and a createek/createak TPM2 quote with the recorded
nonce as a magic-delimited stream (no dd/wc/awk in the initramfs).
scripts/test-measured-capture.py gates the artifacts; the committed
fixture is validated durably by the Rust triangle: the ogir-bootlog
replay equals the live PCR read (PCR 11 included), the quote's
pcrDigest equals SHA256 of the concatenated replayed values and
echoes the nonce, and the QuoteVerifier cryptographically verifies
the signature (tamper rejects).

The multi-bank log also forced a production parser fix: the TCG EFI
Spec ID digestSize field is u16, not u8 - single-bank logs escaped
the bug because the algorithm-table tail is ignored.


* fix: canonicalize the M4-028c planning issue specification

The CI bootstrap validates every planning/issues/*.md against the
live issue tracker: the file needs the labels and milestone header
comments, the H1 must equal the live issue title, and the live body
must be byte-identical to the file. Align the spec (and issue #63's
body, already edited to match byte-for-byte modulo the trailing
newline GitHub appends).


* fix: derive the capture gate's directory from the script location

CodeQL flags the argv-derived artifacts directory as user-controlled
path flow (4 high, py/path-injection family). The gate's purpose is
to validate THE canonical capture output, so the directory is now a
build-controlled constant derived from __file__ - no user input
reaches any file access. Verified against the executed capture.


---------

+++ crates/ogir-bootlog/src/parser.rs
@@ -137,11 +137,13 @@ pub fn parse(bytes: &[u8]) -> Result<EventLog, BootlogError> {
     let mut algorithms = Vec::with_capacity(algorithm_count as usize);
     for _ in 0..algorithm_count {
         let algorithm_id = tail.u16()?;
-        let digest_size = tail.u8()?;
+        // digestSize is a u16 field (TCG EFI Protocol Specification);
+        // multi-bank logs (SHA-1/256/384/512) misalign if read as u8.
+        let digest_size = tail.u16()?;
         if digest_size == 0 || digest_size > 64 {
             return Err(BootlogError::Malformed("digest size out of range"));
         }
-        algorithms.push((algorithm_id, digest_size));
+        algorithms.push((algorithm_id, digest_size as u8));
     }
     // The remaining vendorInfo length byte plus vendor bytes are
     // present in some logs; remaining tail is ignored deliberately
